LangChain_PDF_Groq_Ollama_Embeddings/requirements.txt
streamlit>=1.37.0
python-dotenv>=1.0.1
langchain>=0.2.10
langchain-openai>=0.2.0
langchain-core>=0.2.22
langchain-community>=0.2.9
langchain-groq>=0.2.0
pypdf>=3.10.0
faiss-cpu>=1.8.1